Fix benchmark build errors - cast uint indices to int

+++ MonoGame/Atomic.Net.MonoGame.Benchmarks/Core/SparseArrayBenchmark.cs
@@ -57,7 +57,7 @@ public class SparseArrayBenchmark
         foreach(var (i, value) in _sparseArray)
         {
             result += value;
-            result += i;
+            result += (int)i;
         }
 
         return result;
